+++ CHANGELOG.md
@@ -6,6 +6,14 @@
 
 - Internals: The ability to store and change "naked" properties
 
+### Changed
+
+- Internals: QuoteTemplate and ClozeTemplate use single-value containers
+
+### Removed
+
+- CardTemplateSerializationWrapper and related tests
+
 ## [0.6.2] - 2019-08-29
 
 ### Fixed